+++ Modules/Inventario/Entities/Origen.php
@@ -14,5 +14,10 @@ class Origen extends Model
     protected $fillable = [
         'nombre',
         'descripcion',
+        'activo',
+    ];
+
+    protected $casts = [
+        'activo' => 'boolean',
     ];
 }