+++ config/versiones.php
@@ -1,10 +1,10 @@
 <?php
 
 return [
-    'IEE'           => '1.19.0',
-    'BhagamApps'    => '1.18.0',
+    'IEE'           => '1.20.0',
+    'BhagamApps'    => '1.19.0',
     'User'          => '2.4.4',
-    'Inventario'    => '2.14.0',
+    'Inventario'    => '2.15.0',
     'Apps'          => '1.5.1',
     'CrudGenerator' => '1.1.0',
 ];